requirements.txt
fastapi>=0.109.0
pydantic>=2.6.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0
pillow>=10.2.0
pydantic-settings>=2.0.0
scrapegraph-py>=1.0.0
numpy>=1.26.0
websockets>=12.0
scrapegraphai==1.39.0  # Latest version compatible with Python 3.12
aiohttp>=3.9.0
asyncio>=3.4.3
